% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICS2014Ch.R
\docType{data}
\name{MICS2014Ch}
\alias{MICS2014Ch}
\title{Balochistan Province data from Pakistan Population Census 2017}
\format{A data frame with 31083 rows and 270 variables:
\describe{
  \item{\code{HH1}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH2}}{integer COLUMN_DESCRIPTION}
  \item{\code{LN}}{integer COLUMN_DESCRIPTION}
  \item{\code{UF1}}{integer COLUMN_DESCRIPTION}
  \item{\code{UF2}}{integer COLUMN_DESCRIPTION}
  \item{\code{UF4}}{integer COLUMN_DESCRIPTION}
  \item{\code{UF6}}{integer COLUMN_DESCRIPTION}
  \item{\code{UF7}}{integer COLUMN_DESCRIPTION}
  \item{\code{UF8D}}{integer COLUMN_DESCRIPTION}
  \item{\code{UF8M}}{integer COLUMN_DESCRIPTION}
  \item{\code{UF8Y}}{integer COLUMN_DESCRIPTION}
  \item{\code{UF9}}{integer COLUMN_DESCRIPTION}
  \item{\code{UF10}}{integer COLUMN_DESCRIPTION}
  \item{\code{UF11}}{integer COLUMN_DESCRIPTION}
  \item{\code{UF12H}}{integer COLUMN_DESCRIPTION}
  \item{\code{UF12M}}{integer COLUMN_DESCRIPTION}
  \item{\code{UF13H}}{integer COLUMN_DESCRIPTION}
  \item{\code{UF13M}}{integer COLUMN_DESCRIPTION}
  \item{\code{AG1D}}{integer COLUMN_DESCRIPTION}
  \item{\code{AG1M}}{integer COLUMN_DESCRIPTION}
  \item{\code{AG1Y}}{integer COLUMN_DESCRIPTION}
  \item{\code{AG2}}{integer COLUMN_DESCRIPTION}
  \item{\code{BR1}}{integer COLUMN_DESCRIPTION}
  \item{\code{BR2}}{integer COLUMN_DESCRIPTION}
  \item{\code{BR3}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC1}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC2A}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC2B}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC2C}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC3A}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC3B}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC5}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7AA}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7AB}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7AX}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7AY}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7BA}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7BB}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7BX}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7BY}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7CA}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7CB}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7CX}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7CY}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7DA}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7DB}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7DX}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7DY}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7EA}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7EB}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7EX}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7EY}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7FA}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7FB}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7FX}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC7FY}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC8}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC9}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC10}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC11}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC12}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC13}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC14}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC15}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC16}}{integer COLUMN_DESCRIPTION}
  \item{\code{EC17}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD2}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD3}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD4}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD5}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD6}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD7A}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD7B}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD7C}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD7D}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD7DN}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD7E}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD7EN}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD7F}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD8A}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD8AN}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD8B}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD8C}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD8D}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD8E}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD8F}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD8G}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD8H}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD8I}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD8J}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD8K}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD8L}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD8M}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD8N}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD8O}}{integer COLUMN_DESCRIPTION}
  \item{\code{BD11}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM1}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM2}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3BD}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3BM}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3BY}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3P0D}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3P0M}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3P0Y}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3P1D}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3P1M}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3P1Y}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3P2D}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3P2M}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3P2Y}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3P3D}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3P3M}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3P3Y}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3T1D}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3T1M}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3T1Y}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3T2D}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3T2M}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3T2Y}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3T3D}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3T3M}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3T3Y}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3M1D}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3M1M}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3M1Y}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3M2D}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3M2M}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM3M2Y}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM5}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM6}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM7}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM8}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM9}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM10}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM11}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM12}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM15A}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM15B}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM16}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM19A}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM19B}}{integer COLUMN_DESCRIPTION}
  \item{\code{IM19C}}{integer COLUMN_DESCRIPTION}
  \item{\code{VS1}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA1}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA2}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA3}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA3A}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA3BA}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA3BB}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA3BC}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA3BD}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA3BE}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA3BH}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA3BI}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA3BJ}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA3BK}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA3BL}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA3BO}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA3BP}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA3BQ}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA3BR}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA3BX}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA4A}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA4B}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA4BB}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA4CA}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA4CB}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA4E}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA4FA}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA4FB}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA5}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA6A}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA6B}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA6G}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA6H}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA6L}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA6M}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA6N}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA6O}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA6Q}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA6X}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA6AA}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA6BB}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA7}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA8}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA9}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA10}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA11A}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA11B}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA11C}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA11D}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA11E}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA11H}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA11I}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA11J}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA11K}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA11L}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA11O}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA11P}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA11Q}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA11R}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA11X}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA12}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA13A}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA13B}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA13C}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA13D}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA13E}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA13H}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA13I}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA13J}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA13P}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA13Q}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA13R}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA13X}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA13Z}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA13BB}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA13DD}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA13EE}}{integer COLUMN_DESCRIPTION}
  \item{\code{CA15}}{integer COLUMN_DESCRIPTION}
  \item{\code{AN1}}{integer COLUMN_DESCRIPTION}
  \item{\code{AN2}}{integer COLUMN_DESCRIPTION}
  \item{\code{AN3}}{integer COLUMN_DESCRIPTION}
  \item{\code{AN3A}}{integer COLUMN_DESCRIPTION}
  \item{\code{AN4}}{integer COLUMN_DESCRIPTION}
  \item{\code{AN4A}}{integer COLUMN_DESCRIPTION}
  \item{\code{HAP}}{double COLUMN_DESCRIPTION}
  \item{\code{HAZ}}{double COLUMN_DESCRIPTION}
  \item{\code{HAM}}{double COLUMN_DESCRIPTION}
  \item{\code{WAP}}{double COLUMN_DESCRIPTION}
  \item{\code{WAZ}}{double COLUMN_DESCRIPTION}
  \item{\code{WAM}}{double COLUMN_DESCRIPTION}
  \item{\code{WHP}}{double COLUMN_DESCRIPTION}
  \item{\code{WHZ}}{double COLUMN_DESCRIPTION}
  \item{\code{WHM}}{double COLUMN_DESCRIPTION}
  \item{\code{FLAG}}{integer COLUMN_DESCRIPTION}
  \item{\code{BMI}}{integer COLUMN_DESCRIPTION}
  \item{\code{ZBMI}}{integer COLUMN_DESCRIPTION}
  \item{\code{HAZ2}}{integer COLUMN_DESCRIPTION}
  \item{\code{WAZ2}}{integer COLUMN_DESCRIPTION}
  \item{\code{WHZ2}}{integer COLUMN_DESCRIPTION}
  \item{\code{HAZFLAG}}{integer COLUMN_DESCRIPTION}
  \item{\code{WAZFLAG}}{integer COLUMN_DESCRIPTION}
  \item{\code{WHZFLAG}}{integer COLUMN_DESCRIPTION}
  \item{\code{BMIFLAG}}{integer COLUMN_DESCRIPTION}
  \item{\code{WHZNOAGE}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH6}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH7}}{integer COLUMN_DESCRIPTION}
  \item{\code{HL4}}{integer COLUMN_DESCRIPTION}
  \item{\code{ED4A}}{integer COLUMN_DESCRIPTION}
  \item{\code{ED4B}}{integer COLUMN_DESCRIPTION}
  \item{\code{CDOI}}{integer COLUMN_DESCRIPTION}
  \item{\code{CDOB}}{integer COLUMN_DESCRIPTION}
  \item{\code{CAGE}}{integer COLUMN_DESCRIPTION}
  \item{\code{CAGE.6}}{integer COLUMN_DESCRIPTION}
  \item{\code{CAGE.11}}{integer COLUMN_DESCRIPTION}
  \item{\code{CAGED}}{integer COLUMN_DESCRIPTION}
  \item{\code{melevel}}{integer COLUMN_DESCRIPTION}
  \item{\code{division}}{integer COLUMN_DESCRIPTION}
  \item{\code{hh7r}}{integer COLUMN_DESCRIPTION}
  \item{\code{hh6r}}{integer COLUMN_DESCRIPTION}
  \item{\code{suburban}}{integer COLUMN_DESCRIPTION}
  \item{\code{stratum}}{integer COLUMN_DESCRIPTION}
  \item{\code{wscore}}{double COLUMN_DESCRIPTION}
  \item{\code{windex5}}{integer COLUMN_DESCRIPTION}
  \item{\code{wscoreu}}{double COLUMN_DESCRIPTION}
  \item{\code{windex5u}}{integer COLUMN_DESCRIPTION}
  \item{\code{wscorer}}{double COLUMN_DESCRIPTION}
  \item{\code{windex5r}}{integer COLUMN_DESCRIPTION}
  \item{\code{chweight}}{double COLUMN_DESCRIPTION}
  }}
\usage{
data(MICS2014Ch)
}
\description{
\code{MICS2014Ch} Balochistan Province data from Pakistan Population Census 2017.
}
\examples{
data(MICS2014Ch)
}
\seealso{
\code{\link{MICS2014HH}}
 , \code{\link{MICS2014HL}}
 , \code{\link{MICS2014Wm}}
}
\author{
\enumerate{
  \item{Muhammad Yaseen } {(\email{myaseen208@gmail.com})}
  \item{Muhammad Usman } {(\email{usmann75@hotmail.com})}
  }
}
\keyword{datasets}
