% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICS2014HH.R
\docType{data}
\name{MICS2014HH}
\alias{MICS2014HH}
\title{Balochistan Province data from Pakistan Population Census 2017}
\format{A data frame with 41413 rows and 188 variables:
\describe{
  \item{\code{HH1}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH2}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH3}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH4}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH5D}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH5M}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH5Y}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH6}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH7}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH8A}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH9}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH10}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH11}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH12}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH13}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH14}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH15}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH16}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH17}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH18H}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH18M}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH19H}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH19M}}{integer COLUMN_DESCRIPTION}
  \item{\code{SL1}}{integer COLUMN_DESCRIPTION}
  \item{\code{SL9A}}{integer COLUMN_DESCRIPTION}
  \item{\code{SL9B}}{integer COLUMN_DESCRIPTION}
  \item{\code{SL9C}}{integer COLUMN_DESCRIPTION}
  \item{\code{CD.FLAG}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL2A}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL2B}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL2C}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL2D}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL4}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL5}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL6}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL7A}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL7B}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL7C}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL7D}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL7E}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL7F}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL8}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL9}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL10A}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL10B}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL10C}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL10D}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL10E}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL10F}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL10G}}{integer COLUMN_DESCRIPTION}
  \item{\code{CL12}}{integer COLUMN_DESCRIPTION}
  \item{\code{CD2}}{integer COLUMN_DESCRIPTION}
  \item{\code{CD3A}}{integer COLUMN_DESCRIPTION}
  \item{\code{CD3B}}{integer COLUMN_DESCRIPTION}
  \item{\code{CD3C}}{integer COLUMN_DESCRIPTION}
  \item{\code{CD3D}}{integer COLUMN_DESCRIPTION}
  \item{\code{CD3E}}{integer COLUMN_DESCRIPTION}
  \item{\code{CD3F}}{integer COLUMN_DESCRIPTION}
  \item{\code{CD3G}}{integer COLUMN_DESCRIPTION}
  \item{\code{CD3H}}{integer COLUMN_DESCRIPTION}
  \item{\code{CD3I}}{integer COLUMN_DESCRIPTION}
  \item{\code{CD3J}}{integer COLUMN_DESCRIPTION}
  \item{\code{CD3K}}{integer COLUMN_DESCRIPTION}
  \item{\code{CD4}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC1B}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC2}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC3}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC4}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC5}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC6}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC7}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC8A}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC8B}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC8C}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC8D}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC8E}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC8F}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC8G}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC8H}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC8I}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC8J}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC8K}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC8L}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC8M}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC8N}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC8O}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC9A}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC9B}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC9C}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC9D}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC9E}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC9F}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC9G}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC9H}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC9I}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC10}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC11}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC12}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC13}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC14A}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC14B}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC14C}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC14D}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC14E}}{integer COLUMN_DESCRIPTION}
  \item{\code{HC15}}{integer COLUMN_DESCRIPTION}
  \item{\code{WS1}}{integer COLUMN_DESCRIPTION}
  \item{\code{WS2}}{integer COLUMN_DESCRIPTION}
  \item{\code{WS3}}{integer COLUMN_DESCRIPTION}
  \item{\code{WS4}}{integer COLUMN_DESCRIPTION}
  \item{\code{WS5}}{integer COLUMN_DESCRIPTION}
  \item{\code{WS6}}{integer COLUMN_DESCRIPTION}
  \item{\code{WS7A}}{integer COLUMN_DESCRIPTION}
  \item{\code{WS7B}}{integer COLUMN_DESCRIPTION}
  \item{\code{WS7C}}{integer COLUMN_DESCRIPTION}
  \item{\code{WS7D}}{integer COLUMN_DESCRIPTION}
  \item{\code{WS7E}}{integer COLUMN_DESCRIPTION}
  \item{\code{WS7F}}{integer COLUMN_DESCRIPTION}
  \item{\code{WS7X}}{integer COLUMN_DESCRIPTION}
  \item{\code{WS7Z}}{integer COLUMN_DESCRIPTION}
  \item{\code{WS8}}{integer COLUMN_DESCRIPTION}
  \item{\code{WS9}}{integer COLUMN_DESCRIPTION}
  \item{\code{WS10}}{integer COLUMN_DESCRIPTION}
  \item{\code{WS11}}{integer COLUMN_DESCRIPTION}
  \item{\code{RM1}}{integer COLUMN_DESCRIPTION}
  \item{\code{RM2}}{integer COLUMN_DESCRIPTION}
  \item{\code{RM3A}}{integer COLUMN_DESCRIPTION}
  \item{\code{RM3B}}{integer COLUMN_DESCRIPTION}
  \item{\code{RM3C}}{integer COLUMN_DESCRIPTION}
  \item{\code{RM3D}}{integer COLUMN_DESCRIPTION}
  \item{\code{RM3Z}}{integer COLUMN_DESCRIPTION}
  \item{\code{RM4}}{integer COLUMN_DESCRIPTION}
  \item{\code{RM5}}{integer COLUMN_DESCRIPTION}
  \item{\code{RM5A}}{integer COLUMN_DESCRIPTION}
  \item{\code{PB1}}{integer COLUMN_DESCRIPTION}
  \item{\code{PB2A}}{integer COLUMN_DESCRIPTION}
  \item{\code{PB2B}}{integer COLUMN_DESCRIPTION}
  \item{\code{PB2X}}{integer COLUMN_DESCRIPTION}
  \item{\code{PB2Z}}{integer COLUMN_DESCRIPTION}
  \item{\code{SN1}}{integer COLUMN_DESCRIPTION}
  \item{\code{SN2A}}{integer COLUMN_DESCRIPTION}
  \item{\code{SN2B}}{integer COLUMN_DESCRIPTION}
  \item{\code{SN2C}}{integer COLUMN_DESCRIPTION}
  \item{\code{SN2D}}{integer COLUMN_DESCRIPTION}
  \item{\code{SN2E}}{integer COLUMN_DESCRIPTION}
  \item{\code{SN2X}}{integer COLUMN_DESCRIPTION}
  \item{\code{SN2Z}}{integer COLUMN_DESCRIPTION}
  \item{\code{SN3}}{integer COLUMN_DESCRIPTION}
  \item{\code{SN4}}{integer COLUMN_DESCRIPTION}
  \item{\code{SN6}}{integer COLUMN_DESCRIPTION}
  \item{\code{SN7}}{integer COLUMN_DESCRIPTION}
  \item{\code{SN8}}{integer COLUMN_DESCRIPTION}
  \item{\code{HW1}}{integer COLUMN_DESCRIPTION}
  \item{\code{HW2}}{integer COLUMN_DESCRIPTION}
  \item{\code{HW3A}}{integer COLUMN_DESCRIPTION}
  \item{\code{HW3BA}}{integer COLUMN_DESCRIPTION}
  \item{\code{HW3BB}}{integer COLUMN_DESCRIPTION}
  \item{\code{HW3BC}}{integer COLUMN_DESCRIPTION}
  \item{\code{HW3BD}}{integer COLUMN_DESCRIPTION}
  \item{\code{HW4}}{integer COLUMN_DESCRIPTION}
  \item{\code{HW5A}}{integer COLUMN_DESCRIPTION}
  \item{\code{HW5BA}}{integer COLUMN_DESCRIPTION}
  \item{\code{HW5BB}}{integer COLUMN_DESCRIPTION}
  \item{\code{HW5BC}}{integer COLUMN_DESCRIPTION}
  \item{\code{HW5BD}}{integer COLUMN_DESCRIPTION}
  \item{\code{SI1}}{integer COLUMN_DESCRIPTION}
  \item{\code{SI3}}{integer COLUMN_DESCRIPTION}
  \item{\code{SI4}}{integer COLUMN_DESCRIPTION}
  \item{\code{SI5}}{integer COLUMN_DESCRIPTION}
  \item{\code{SI6}}{integer COLUMN_DESCRIPTION}
  \item{\code{SI7}}{integer COLUMN_DESCRIPTION}
  \item{\code{SI8CN}}{integer COLUMN_DESCRIPTION}
  \item{\code{SI8HH}}{integer COLUMN_DESCRIPTION}
  \item{\code{HHSEX}}{integer COLUMN_DESCRIPTION}
  \item{\code{division}}{integer COLUMN_DESCRIPTION}
  \item{\code{hh7r}}{integer COLUMN_DESCRIPTION}
  \item{\code{hh6r}}{integer COLUMN_DESCRIPTION}
  \item{\code{suburban}}{integer COLUMN_DESCRIPTION}
  \item{\code{SIRes}}{integer COLUMN_DESCRIPTION}
  \item{\code{stratum}}{integer COLUMN_DESCRIPTION}
  \item{\code{helevel}}{integer COLUMN_DESCRIPTION}
  \item{\code{wscore}}{double COLUMN_DESCRIPTION}
  \item{\code{windex5}}{integer COLUMN_DESCRIPTION}
  \item{\code{wscoreu}}{double COLUMN_DESCRIPTION}
  \item{\code{windex5u}}{integer COLUMN_DESCRIPTION}
  \item{\code{wscorer}}{double COLUMN_DESCRIPTION}
  \item{\code{windex5r}}{integer COLUMN_DESCRIPTION}
  \item{\code{hhweight}}{double COLUMN_DESCRIPTION}
  \item{\code{stweight}}{double COLUMN_DESCRIPTION} 
  }}
\usage{
data(MICS2014HH)
}
\description{
\code{MICS2014HH} Balochistan Province data from Pakistan Population Census 2017.
}
\examples{
data(MICS2014HH)
}
\seealso{
\code{\link{MICS2014Ch}}
 , \code{\link{MICS2014HL}}
 , \code{\link{MICS2014Wm}}
}
\author{
\enumerate{
  \item{Muhammad Yaseen } {(\email{myaseen208@gmail.com})}
  \item{Muhammad Usman } {(\email{usmann75@hotmail.com})}
  }
}
\keyword{datasets}
