% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICS2014Wm.R
\docType{data}
\name{MICS2014Wm}
\alias{MICS2014Wm}
\title{Balochistan Province data from Pakistan Population Census 2017}
\format{A data frame with 61286 rows and 308 variables:
\describe{
  \item{\code{HH1}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH2}}{integer COLUMN_DESCRIPTION}
  \item{\code{LN}}{integer COLUMN_DESCRIPTION}
  \item{\code{WM1}}{integer COLUMN_DESCRIPTION}
  \item{\code{WM2}}{integer COLUMN_DESCRIPTION}
  \item{\code{WM4}}{integer COLUMN_DESCRIPTION}
  \item{\code{WM5}}{integer COLUMN_DESCRIPTION}
  \item{\code{WM6D}}{integer COLUMN_DESCRIPTION}
  \item{\code{WM6M}}{integer COLUMN_DESCRIPTION}
  \item{\code{WM6Y}}{integer COLUMN_DESCRIPTION}
  \item{\code{WM7}}{integer COLUMN_DESCRIPTION}
  \item{\code{WM8}}{integer COLUMN_DESCRIPTION}
  \item{\code{WM9}}{integer COLUMN_DESCRIPTION}
  \item{\code{WM10H}}{integer COLUMN_DESCRIPTION}
  \item{\code{WM10M}}{integer COLUMN_DESCRIPTION}
  \item{\code{WM11H}}{integer COLUMN_DESCRIPTION}
  \item{\code{WM11M}}{integer COLUMN_DESCRIPTION}
  \item{\code{WB1M}}{integer COLUMN_DESCRIPTION}
  \item{\code{WB1Y}}{integer COLUMN_DESCRIPTION}
  \item{\code{WB2}}{integer COLUMN_DESCRIPTION}
  \item{\code{WB3}}{integer COLUMN_DESCRIPTION}
  \item{\code{WB4}}{integer COLUMN_DESCRIPTION}
  \item{\code{WB5}}{integer COLUMN_DESCRIPTION}
  \item{\code{WB7}}{integer COLUMN_DESCRIPTION}
  \item{\code{WDOI}}{integer COLUMN_DESCRIPTION}
  \item{\code{WB1C}}{integer COLUMN_DESCRIPTION}
  \item{\code{WB1F}}{integer COLUMN_DESCRIPTION}
  \item{\code{WB2C}}{integer COLUMN_DESCRIPTION}
  \item{\code{WAGE}}{integer COLUMN_DESCRIPTION}
  \item{\code{MT2}}{integer COLUMN_DESCRIPTION}
  \item{\code{MT3}}{integer COLUMN_DESCRIPTION}
  \item{\code{MT4}}{integer COLUMN_DESCRIPTION}
  \item{\code{MT6}}{integer COLUMN_DESCRIPTION}
  \item{\code{MT7}}{integer COLUMN_DESCRIPTION}
  \item{\code{MT8}}{integer COLUMN_DESCRIPTION}
  \item{\code{MT9}}{integer COLUMN_DESCRIPTION}
  \item{\code{MT10}}{integer COLUMN_DESCRIPTION}
  \item{\code{MT11}}{integer COLUMN_DESCRIPTION}
  \item{\code{MA1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MA2}}{integer COLUMN_DESCRIPTION}
  \item{\code{MA3}}{integer COLUMN_DESCRIPTION}
  \item{\code{MA4}}{integer COLUMN_DESCRIPTION}
  \item{\code{MA5}}{integer COLUMN_DESCRIPTION}
  \item{\code{MA6}}{integer COLUMN_DESCRIPTION}
  \item{\code{MA8M}}{integer COLUMN_DESCRIPTION}
  \item{\code{MA8Y}}{integer COLUMN_DESCRIPTION}
  \item{\code{MA9}}{integer COLUMN_DESCRIPTION}
  \item{\code{MA8C}}{integer COLUMN_DESCRIPTION}
  \item{\code{MA8F}}{integer COLUMN_DESCRIPTION}
  \item{\code{MA9C}}{integer COLUMN_DESCRIPTION}
  \item{\code{CM1}}{integer COLUMN_DESCRIPTION}
  \item{\code{CM2M}}{integer COLUMN_DESCRIPTION}
  \item{\code{CM2Y}}{integer COLUMN_DESCRIPTION}
  \item{\code{CM3}}{integer COLUMN_DESCRIPTION}
  \item{\code{CM4}}{integer COLUMN_DESCRIPTION}
  \item{\code{CM5A}}{integer COLUMN_DESCRIPTION}
  \item{\code{CM5B}}{integer COLUMN_DESCRIPTION}
  \item{\code{CM6}}{integer COLUMN_DESCRIPTION}
  \item{\code{CM7A}}{integer COLUMN_DESCRIPTION}
  \item{\code{CM7B}}{integer COLUMN_DESCRIPTION}
  \item{\code{CM8}}{integer COLUMN_DESCRIPTION}
  \item{\code{CM9A}}{integer COLUMN_DESCRIPTION}
  \item{\code{CM9B}}{integer COLUMN_DESCRIPTION}
  \item{\code{CM10}}{integer COLUMN_DESCRIPTION}
  \item{\code{CTOT}}{integer COLUMN_DESCRIPTION}
  \item{\code{CM12M}}{integer COLUMN_DESCRIPTION}
  \item{\code{CM12Y}}{integer COLUMN_DESCRIPTION}
  \item{\code{CM13}}{integer COLUMN_DESCRIPTION}
  \item{\code{DB1}}{integer COLUMN_DESCRIPTION}
  \item{\code{DB2}}{integer COLUMN_DESCRIPTION}
  \item{\code{DB3U}}{integer COLUMN_DESCRIPTION}
  \item{\code{DB3N}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN2A}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN2B}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN2D}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN2E}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN2F}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN2H}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN2X}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN2AU}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN2AN}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN3}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN4A}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN4B}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN4C}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN4D}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN5}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN6}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN7}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN9}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN10}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN11}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN13}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN14A}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN14B}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN14X}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN14Z}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN16}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN17A}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN17B}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN17D}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN17F}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN17H}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN17X}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN17Y}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN18}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN19}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN19A}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN20}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN21}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN22A}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN22}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN23}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN24}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN25U}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN25N}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN26}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN27A}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN27B}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN27C}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN27D}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN27E}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN27F}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN27G}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN27H}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN27I}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN27J}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN27X}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN28}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN29A}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN29B}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN29C}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN29X}}{integer COLUMN_DESCRIPTION}
  \item{\code{MN29Z}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN2U}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN2N}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN3}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN4}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN5}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN7}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN8}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN9}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN10}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN11}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN12U}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN12N}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN13A}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN13B}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN13D}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN13E}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN13F}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN13H}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN13X}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN14}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN16}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN18}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN19}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN20}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN21U}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN21N}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN22A}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN22B}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN22D}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN22E}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN22F}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN22H}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN22X}}{integer COLUMN_DESCRIPTION}
  \item{\code{PN23}}{integer COLUMN_DESCRIPTION}
  \item{\code{IS2A}}{integer COLUMN_DESCRIPTION}
  \item{\code{IS2B}}{integer COLUMN_DESCRIPTION}
  \item{\code{IS2C}}{integer COLUMN_DESCRIPTION}
  \item{\code{IS2D}}{integer COLUMN_DESCRIPTION}
  \item{\code{IS2E}}{integer COLUMN_DESCRIPTION}
  \item{\code{IS2F}}{integer COLUMN_DESCRIPTION}
  \item{\code{IS2G}}{integer COLUMN_DESCRIPTION}
  \item{\code{IS2H}}{integer COLUMN_DESCRIPTION}
  \item{\code{IS2X}}{integer COLUMN_DESCRIPTION}
  \item{\code{IS2Y}}{integer COLUMN_DESCRIPTION}
  \item{\code{IS2Z}}{integer COLUMN_DESCRIPTION}
  \item{\code{CP1}}{integer COLUMN_DESCRIPTION}
  \item{\code{CP2}}{integer COLUMN_DESCRIPTION}
  \item{\code{CP2A}}{integer COLUMN_DESCRIPTION}
  \item{\code{CP3A}}{integer COLUMN_DESCRIPTION}
  \item{\code{CP3B}}{integer COLUMN_DESCRIPTION}
  \item{\code{CP3C}}{integer COLUMN_DESCRIPTION}
  \item{\code{CP3D}}{integer COLUMN_DESCRIPTION}
  \item{\code{CP3E}}{integer COLUMN_DESCRIPTION}
  \item{\code{CP3F}}{integer COLUMN_DESCRIPTION}
  \item{\code{CP3G}}{integer COLUMN_DESCRIPTION}
  \item{\code{CP3H}}{integer COLUMN_DESCRIPTION}
  \item{\code{CP3I}}{integer COLUMN_DESCRIPTION}
  \item{\code{CP3L}}{integer COLUMN_DESCRIPTION}
  \item{\code{CP3M}}{integer COLUMN_DESCRIPTION}
  \item{\code{CP3X}}{integer COLUMN_DESCRIPTION}
  \item{\code{UN2}}{integer COLUMN_DESCRIPTION}
  \item{\code{UN3}}{integer COLUMN_DESCRIPTION}
  \item{\code{UN4}}{integer COLUMN_DESCRIPTION}
  \item{\code{UN6}}{integer COLUMN_DESCRIPTION}
  \item{\code{UN7U}}{integer COLUMN_DESCRIPTION}
  \item{\code{UN7N}}{integer COLUMN_DESCRIPTION}
  \item{\code{UN10}}{integer COLUMN_DESCRIPTION}
  \item{\code{UN11A}}{integer COLUMN_DESCRIPTION}
  \item{\code{UN11B}}{integer COLUMN_DESCRIPTION}
  \item{\code{UN11C}}{integer COLUMN_DESCRIPTION}
  \item{\code{UN11D}}{integer COLUMN_DESCRIPTION}
  \item{\code{UN11E}}{integer COLUMN_DESCRIPTION}
  \item{\code{UN11F}}{integer COLUMN_DESCRIPTION}
  \item{\code{UN11G}}{integer COLUMN_DESCRIPTION}
  \item{\code{UN11H}}{integer COLUMN_DESCRIPTION}
  \item{\code{UN11I}}{integer COLUMN_DESCRIPTION}
  \item{\code{UN11X}}{integer COLUMN_DESCRIPTION}
  \item{\code{UN11Z}}{integer COLUMN_DESCRIPTION}
  \item{\code{UN13U}}{integer COLUMN_DESCRIPTION}
  \item{\code{UN13N}}{integer COLUMN_DESCRIPTION}
  \item{\code{DV1A}}{integer COLUMN_DESCRIPTION}
  \item{\code{DV1B}}{integer COLUMN_DESCRIPTION}
  \item{\code{DV1C}}{integer COLUMN_DESCRIPTION}
  \item{\code{DV1D}}{integer COLUMN_DESCRIPTION}
  \item{\code{DV1E}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA1}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA2}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA3}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA4}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA5}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA6}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA7}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA8A}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA8B}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA8C}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA9}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA10}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA11}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA12}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA15A}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA15B}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA15C}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA15D}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA16}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA17}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA18}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA20}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA21}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA22}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA23}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA24}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA25}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA26}}{integer COLUMN_DESCRIPTION}
  \item{\code{HA27}}{integer COLUMN_DESCRIPTION}
  \item{\code{TA1}}{integer COLUMN_DESCRIPTION}
  \item{\code{TA2}}{integer COLUMN_DESCRIPTION}
  \item{\code{TA3}}{integer COLUMN_DESCRIPTION}
  \item{\code{TA4}}{integer COLUMN_DESCRIPTION}
  \item{\code{TA5}}{integer COLUMN_DESCRIPTION}
  \item{\code{TA6}}{integer COLUMN_DESCRIPTION}
  \item{\code{TA7}}{integer COLUMN_DESCRIPTION}
  \item{\code{TA8A}}{integer COLUMN_DESCRIPTION}
  \item{\code{TA8B}}{integer COLUMN_DESCRIPTION}
  \item{\code{TA8C}}{integer COLUMN_DESCRIPTION}
  \item{\code{TA8D}}{integer COLUMN_DESCRIPTION}
  \item{\code{TA8X}}{integer COLUMN_DESCRIPTION}
  \item{\code{TA9}}{integer COLUMN_DESCRIPTION}
  \item{\code{TA10}}{integer COLUMN_DESCRIPTION}
  \item{\code{TA11}}{integer COLUMN_DESCRIPTION}
  \item{\code{TA12A}}{integer COLUMN_DESCRIPTION}
  \item{\code{TA12D}}{integer COLUMN_DESCRIPTION}
  \item{\code{TA12E}}{integer COLUMN_DESCRIPTION}
  \item{\code{TA12F}}{integer COLUMN_DESCRIPTION}
  \item{\code{TA12X}}{integer COLUMN_DESCRIPTION}
  \item{\code{TA13}}{integer COLUMN_DESCRIPTION}
  \item{\code{LS2}}{integer COLUMN_DESCRIPTION}
  \item{\code{LS3}}{integer COLUMN_DESCRIPTION}
  \item{\code{LS4}}{integer COLUMN_DESCRIPTION}
  \item{\code{LS5}}{integer COLUMN_DESCRIPTION}
  \item{\code{LS6}}{integer COLUMN_DESCRIPTION}
  \item{\code{LS7}}{integer COLUMN_DESCRIPTION}
  \item{\code{LS8}}{integer COLUMN_DESCRIPTION}
  \item{\code{LS9}}{integer COLUMN_DESCRIPTION}
  \item{\code{LS10}}{integer COLUMN_DESCRIPTION}
  \item{\code{LS11}}{integer COLUMN_DESCRIPTION}
  \item{\code{LS12}}{integer COLUMN_DESCRIPTION}
  \item{\code{LS13}}{integer COLUMN_DESCRIPTION}
  \item{\code{LS14}}{integer COLUMN_DESCRIPTION}
  \item{\code{LS15}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH6}}{integer COLUMN_DESCRIPTION}
  \item{\code{HH7}}{integer COLUMN_DESCRIPTION}
  \item{\code{WDOB}}{integer COLUMN_DESCRIPTION}
  \item{\code{WDOM}}{integer COLUMN_DESCRIPTION}
  \item{\code{WAGEM}}{integer COLUMN_DESCRIPTION}
  \item{\code{WDOBFC}}{integer COLUMN_DESCRIPTION}
  \item{\code{WDOBLC}}{integer COLUMN_DESCRIPTION}
  \item{\code{MSTATUS}}{integer COLUMN_DESCRIPTION}
  \item{\code{CEB}}{integer COLUMN_DESCRIPTION}
  \item{\code{CSURV}}{integer COLUMN_DESCRIPTION}
  \item{\code{CDEAD}}{integer COLUMN_DESCRIPTION}
  \item{\code{welevel}}{integer COLUMN_DESCRIPTION}
  \item{\code{division}}{integer COLUMN_DESCRIPTION}
  \item{\code{hh7r}}{integer COLUMN_DESCRIPTION}
  \item{\code{hh6r}}{integer COLUMN_DESCRIPTION}
  \item{\code{suburban}}{integer COLUMN_DESCRIPTION}
  \item{\code{stratum}}{integer COLUMN_DESCRIPTION}
  \item{\code{wmweight}}{double COLUMN_DESCRIPTION}
  \item{\code{wscore}}{double COLUMN_DESCRIPTION}
  \item{\code{windex5}}{integer COLUMN_DESCRIPTION}
  \item{\code{wscoreu}}{double COLUMN_DESCRIPTION}
  \item{\code{windex5u}}{integer COLUMN_DESCRIPTION}
  \item{\code{wscorer}}{double COLUMN_DESCRIPTION}
  \item{\code{windex5r}}{integer COLUMN_DESCRIPTION} 
}}
\usage{
data(MICS2014Wm)
}
\description{
\code{MICS2014Wm} Balochistan Province data from Pakistan Population Census 2017.
}
\examples{
data(MICS2014Wm)
}
\seealso{
\code{\link{MICS2014Ch}}
 , \code{\link{MICS2014HH}}
 , \code{\link{MICS2014HL}}
}
\author{
\enumerate{
  \item{Muhammad Yaseen } {(\email{myaseen208@gmail.com})}
  \item{Muhammad Usman } {(\email{usmann75@hotmail.com})}
  }
}
\keyword{datasets}
